--- alembic/versions/08c5bc137b1b_init.py
@@ -1,32 +1,0 @@
-"""init
-
-Revision ID: 08c5bc137b1b
-Revises: 875f1a72ef37
-Create Date: 2023-05-20 10:23:16.937828
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = '08c5bc137b1b'
-down_revision = '875f1a72ef37'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('article', sa.Column('article_title', sa.String(), nullable=False))
-    op.add_column('article', sa.Column('article_author', sa.String(), nullable=True))
-    op.drop_column('article', 'article_authors')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('article', sa.Column('article_authors', postgresql.ARRAY(sa.VARCHAR()), autoincrement=False, nullable=True))
-    op.drop_column('article', 'article_author')
-    op.drop_column('article', 'article_title')
-    # ### end Alembic commands ###
