--- alembic/versions/875f1a72ef37_init.py
@@ -1,58 +1,0 @@
-"""init
-
-Revision ID: 875f1a72ef37
-Revises: f02e8c9f07c1
-Create Date: 2023-05-19 15:32:40.861797
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '875f1a72ef37'
-down_revision = 'f02e8c9f07c1'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('article',
-    sa.Column('id_article', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('article_authors', sa.ARRAY(sa.String()), nullable=True),
-    sa.Column('article_keywords', sa.ARRAY(sa.String()), nullable=True),
-    sa.Column('article_date', sa.Date(), nullable=False),
-    sa.Column('article_text', sa.String(), nullable=False),
-    sa.PrimaryKeyConstraint('id_article')
-    )
-    op.create_table('category',
-    sa.Column('id_category', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('category_name', sa.String(), nullable=False),
-    sa.PrimaryKeyConstraint('id_category'),
-    sa.UniqueConstraint('category_name')
-    )
-    op.create_table('article_category',
-    sa.Column('id_article_category', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('id_article', sa.Integer(), nullable=True),
-    sa.Column('id_category', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['id_article'], ['article.id_article'], onupdate='RESTRICT', ondelete='RESTRICT'),
-    sa.ForeignKeyConstraint(['id_category'], ['category.id_category'], onupdate='RESTRICT', ondelete='RESTRICT'),
-    sa.PrimaryKeyConstraint('id_article_category')
-    )
-    op.drop_table('data')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('data',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('data', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('id', name='data_pkey'),
-    sa.UniqueConstraint('data', name='data_data_key')
-    )
-    op.drop_table('article_category')
-    op.drop_table('category')
-    op.drop_table('article')
-    # ### end Alembic commands ###
