+++ alembic/versions/b92fe8a51809_init.py
@@ -1,0 +1,52 @@
+"""init
+
+Revision ID: b92fe8a51809
+Revises: 
+Create Date: 2023-05-21 06:29:58.439906
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = 'b92fe8a51809'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('article',
+    sa.Column('article_id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('article_title', sa.String(), nullable=False),
+    sa.Column('article_author', sa.String(), nullable=True),
+    sa.Column('article_keywords', sa.ARRAY(sa.String()), nullable=True),
+    sa.Column('article_date', sa.Date(), nullable=False),
+    sa.Column('article_text', sa.String(), nullable=False),
+    sa.PrimaryKeyConstraint('article_id')
+    )
+    op.create_table('category',
+    sa.Column('category_id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('category_name', sa.String(), nullable=False),
+    sa.PrimaryKeyConstraint('category_id'),
+    sa.UniqueConstraint('category_name')
+    )
+    op.create_table('article_category',
+    sa.Column('article_category_id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('article_id', sa.Integer(), nullable=True),
+    sa.Column('category_id', sa.Integer(), nullable=True),
+    sa.ForeignKeyConstraint(['article_id'], ['article.article_id'], onupdate='RESTRICT', ondelete='RESTRICT'),
+    sa.ForeignKeyConstraint(['category_id'], ['category.category_id'], onupdate='RESTRICT', ondelete='RESTRICT'),
+    sa.PrimaryKeyConstraint('article_category_id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('article_category')
+    op.drop_table('category')
+    op.drop_table('article')
+    # ### end Alembic commands ###
