--- alembic/versions/f02e8c9f07c1_init.py
@@ -1,33 +1,0 @@
-"""init
-
-Revision ID: f02e8c9f07c1
-Revises: 
-Create Date: 2023-05-13 11:54:30.534192
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'f02e8c9f07c1'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('data',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('data', sa.String(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('data')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('data')
-    # ### end Alembic commands ###
